--- src/main/java/com/minionz/backend/user/controller/dto/Role.java
@@ -1,5 +1,0 @@
-package com.minionz.backend.user.controller.dto;
-
-public enum Role {
-    OWNER, USER
-}
