+++ src/main/java/com/minionz/backend/user/controller/dto/UserPageResponseDto.java
@@ -1,7 +1,6 @@
 package com.minionz.backend.user.controller.dto;
 
 import com.minionz.backend.common.domain.Address;
-import com.minionz.backend.user.domain.Owner;
 import com.minionz.backend.user.domain.User;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,22 +9,14 @@ import java.util.List;
 
 @Getter
 @NoArgsConstructor
-public class UserPageResponseDto {
+public class UserPageResponseDto extends MyPageResponseDto{
 
-    private String nickname;
-    private String telNumber;
     private Address address;
     private List<UserVisitResponseDto> userVisitResponseList;
 
     public UserPageResponseDto(User user, List<UserVisitResponseDto> userVisitResponseList) {
-        this.nickname = user.getNickName();
+        super(user.getNickName(), user.getTelNumber());
         this.address = user.getAddress();
-        this.telNumber = user.getTelNumber();
         this.userVisitResponseList = userVisitResponseList;
-    }
-
-    public UserPageResponseDto(Owner owner) {
-        this.nickname = owner.getName();
-        this.telNumber = owner.getTelNumber();
     }
 }
