+++ src/test/java/com/minionz/backend/user/controller/UserControllerTest.java
@@ -54,7 +54,7 @@ class UserControllerTest extends ApiDocument {
     @Test
     public void 유저로그인테스트_성공() throws Exception {
         Long id = 1L;
-        final LoginRequestDto LoginRequestDto = new LoginRequestDto("email", "password",Role.USER);
+        final LoginRequestDto LoginRequestDto = new LoginRequestDto("email", "password");
         final LoginResponseDto loginResponseDto = new LoginResponseDto(id, new Message("로그인 성공"));
         willReturn(loginResponseDto).given(userService).login(any(LoginRequestDto.class));
         final ResultActions resultActions = 유저_로그인_요청(LoginRequestDto);
@@ -64,41 +64,22 @@ class UserControllerTest extends ApiDocument {
     @DisplayName("유저 로그인 실패")
     @Test
     public void 유저로그인테스트_실패() throws Exception {
-        final LoginRequestDto LoginRequestDto = new LoginRequestDto("email1", "password",Role.USER);
+        final LoginRequestDto LoginRequestDto = new LoginRequestDto("email1", "password");
         Message errorMessage = new Message("로그인 실패");
         willThrow(new NotFoundException("로그인 실패")).given(userService).login(any(LoginRequestDto.class));
         final ResultActions resultActions = 유저_로그인_요청(LoginRequestDto);
         유저_로그인_실패(errorMessage, resultActions);
     }
 
-    @DisplayName("오너 로그인 성공")
-    @Test
-    public void 오너로그인테스트_성공() throws Exception {
-        Long id = 1L;
-        final LoginRequestDto LoginRequestDto = new LoginRequestDto("email", "password",Role.OWNER);
-        final LoginResponseDto loginResponseDto = new LoginResponseDto(id, new Message("로그인 성공"));
-        willReturn(loginResponseDto).given(userService).login(any(LoginRequestDto.class));
-        final ResultActions resultActions = 유저_로그인_요청(LoginRequestDto);
-        오너_로그인_성공(loginResponseDto, resultActions);
-    }
 
-    @DisplayName("오너 로그인 실패")
-    @Test
-    public void 오너로그인테스트_실패() throws Exception {
-        final LoginRequestDto LoginRequestDto = new LoginRequestDto("email1", "password",Role.OWNER);
-        Message errorMessage = new Message("로그인 실패");
-        willThrow(new NotFoundException("로그인 실패")).given(userService).login(any(LoginRequestDto.class));
-        final ResultActions resultActions = 유저_로그인_요청(LoginRequestDto);
-        오너_로그인_실패(errorMessage, resultActions);
-    }
 
     @DisplayName("유저 로그아웃 성공")
     @Test
     public void 유저로그아웃테스트_성공() throws Exception {
         final Long id = 1L;
         Message message = new Message("로그아웃 성공");
-        willReturn(message).given(userService).logout(any(Long.class), any(Role.class));
-        final ResultActions resultActions = 유저_로그아웃_요청(id, Role.USER);
+        willReturn(message).given(userService).logout(any(Long.class));
+        final ResultActions resultActions = 유저_로그아웃_요청(id);
         유저_로그아웃_성공(resultActions);
     }
 
@@ -107,30 +88,12 @@ class UserControllerTest extends ApiDocument {
     public void 유저로그아웃테스트_실패() throws Exception {
         final Long id = 1L;
         Message errorMessage = new Message("로그아웃 실패");
-        willThrow(new NotFoundException("로그아웃 실패")).given(userService).logout(any(Long.class), any(Role.class));
-        final ResultActions resultActions = 유저_로그아웃_요청(id, Role.USER);
+        willThrow(new NotFoundException("로그아웃 실패")).given(userService).logout(any(Long.class));
+        final ResultActions resultActions = 유저_로그아웃_요청(id);
         유저_로그아웃_실패(errorMessage, resultActions);
     }
 
-    @DisplayName("오너 로그아웃 성공")
-    @Test
-    public void 오너로그아웃테스트_성공() throws Exception {
-        final Long id = 1L;
-        Message message = new Message("로그아웃 성공");
-        willReturn(message).given(userService).logout(any(Long.class), any(Role.class));
-        final ResultActions resultActions = 유저_로그아웃_요청(id, Role.OWNER);
-        오너_로그아웃_성공(resultActions);
-    }
 
-    @DisplayName("오너 로그아웃 실패")
-    @Test
-    public void 오너로그아웃테스트_실패() throws Exception {
-        final Long id = 1L;
-        Message errorMessage = new Message("로그아웃 실패");
-        willThrow(new NotFoundException("로그아웃 실패")).given(userService).logout(any(Long.class), any(Role.class));
-        final ResultActions resultActions = 유저_로그아웃_요청(id, Role.OWNER);
-        오너_로그아웃_실패(errorMessage, resultActions);
-    }
 
     @DisplayName("유저 회원가입 성공")
     @Test
@@ -143,7 +106,6 @@ class UserControllerTest extends ApiDocument {
                 .telNumber("11111111")
                 .password("1234")
                 .address(address)
-                .role(Role.USER)
                 .build();
         Message message = new Message("회원가입 성공");
         willReturn(message).given(userService).signUp(any(JoinRequestDto.class));
@@ -162,45 +124,10 @@ class UserControllerTest extends ApiDocument {
                 .telNumber("11111111")
                 .password("1234")
                 .address(address)
-                .role(Role.USER)
                 .build();Message errorMessage = new Message("회원가입 실패");
         willThrow(new BadRequestException("회원가입 실패")).given(userService).signUp(any(JoinRequestDto.class));
         final ResultActions response = 유저_회원가입_요청(signUpRequest);
         유저_회원가입_실패(errorMessage, response);
-    }
-
-    @DisplayName("오너 회원가입 성공")
-    @Test
-    void 오너회원가입_성공() throws Exception {
-        Address address = new Address("안산시", "상록구", "성포동");
-        JoinRequestDto signUpRequest = JoinRequestDto.builder()
-                .name("정재욱")
-                .email("operation@naver.com")
-                .telNumber("11111111")
-                .password("1234")
-                .role(Role.OWNER)
-                .build();
-        Message message = new Message("회원가입 성공");
-        willReturn(message).given(userService).signUp(any(JoinRequestDto.class));
-        final ResultActions response = 유저_회원가입_요청(signUpRequest);
-        오너_회원가입_성공(message, response);
-    }
-
-    @DisplayName("오너 회원가입 실패")
-    @Test
-    void 오너회원가입_실패() throws Exception {
-        final Address address = new Address("안산시", "상록구", "성포동");
-        JoinRequestDto signUpRequest = JoinRequestDto.builder()
-                .name("정재욱")
-                .email("operation@naver.com")
-                .telNumber("11111111")
-                .password("1234")
-                .role(Role.OWNER)
-                .build();
-        Message errorMessage = new Message("회원가입 실패");
-        willThrow(new BadRequestException("회원가입 실패")).given(userService).signUp(any(JoinRequestDto.class));
-        final ResultActions response = 유저_회원가입_요청(signUpRequest);
-        오너_회원가입_실패(errorMessage, response);
     }
 
     @DisplayName("유저회원탈퇴 성공")
@@ -208,8 +135,8 @@ class UserControllerTest extends ApiDocument {
     void 유저회원탈퇴_성공() throws Exception {
         final Long id = 1L;
         Message message = new Message("회원탈퇴 성공");
-        willReturn(message).given(userService).withdraw(any(Long.class), any(Role.class));
-        final ResultActions response = 유저_회원탈퇴_요청(id, Role.USER);
+        willReturn(message).given(userService).withdraw(any(Long.class));
+        final ResultActions response = 유저_회원탈퇴_요청(id);
         유저_회원탈퇴_성공(response);
     }
 
@@ -218,123 +145,9 @@ class UserControllerTest extends ApiDocument {
     void 유저회원탈퇴_실패() throws Exception {
         final Long id = 1L;
         Message errorMessage = new Message("회원탈퇴 실패");
-        willThrow(new NotFoundException("회원탈퇴 실패")).given(userService).withdraw(any(Long.class), any(Role.class));
-        final ResultActions response = 유저_회원탈퇴_요청(id, Role.USER);
+        willThrow(new NotFoundException("회원탈퇴 실패")).given(userService).withdraw(any(Long.class));
+        final ResultActions response = 유저_회원탈퇴_요청(id);
         유저_회원탈퇴_실패(errorMessage, response);
-    }
-
-    @DisplayName("유저 마이페이지 조회 성공")
-    @Test
-    void 유저마이페이지_조회_성공() throws Exception {
-        Address address = new Address("a", "b", "C");
-        User user = User.builder()
-                .nickName("asd")
-                .telNumber("111")
-                .address(address)
-                .build();
-        Long id = 1L;
-        UserPageResponseDto userPageResponseDto = new UserPageResponseDto(user,null);
-        willReturn(userPageResponseDto).given(userService).viewMypage(any(Long.class), any(Role.class));
-        final ResultActions response = 유저_마이페이지_요청(id, Role.USER);
-        유저_마이페이지_성공(response, userPageResponseDto);
-    }
-
-    @DisplayName("유저 마이페이지 조회 실패")
-    @Test
-    void 유저마이페이지_조회_실패() throws Exception {
-        Long id = 2L;
-        Message errorMessage = new Message("마이페이지 조회 실패");
-        willThrow(new NotFoundException("마이페이지 조회 실패")).given(userService).viewMypage(any(Long.class), any(Role.class));
-        final ResultActions response = 유저_마이페이지_요청(id, Role.USER);
-        유저_마이페이지_실패(response, errorMessage);
-    }
-
-    @DisplayName("오너 마이페이지 조회 성공")
-    @Test
-    void 오너마이페이지_조회_성공() throws Exception {
-        Owner owner = Owner.builder()
-                .telNumber("111")
-                .name("주인")
-                .build();
-        Long id = 1L;
-        UserPageResponseDto userPageResponseDto = new UserPageResponseDto(owner);
-        willReturn(userPageResponseDto).given(userService).viewMypage(any(Long.class), any(Role.class));
-        final ResultActions response = 유저_마이페이지_요청(id, Role.OWNER);
-        오너_마이페이지_성공(response, userPageResponseDto);
-    }
-
-    @DisplayName("오너 마이페이지 조회 실패")
-    @Test
-    void 오너마이페이지_조회_실패() throws Exception {
-        Long id = 2L;
-        Message errorMessage = new Message("마이페이지 조회 실패");
-        willThrow(new NotFoundException("마이페이지 조회 실패")).given(userService).viewMypage(any(Long.class), any(Role.class));
-        final ResultActions response = 유저_마이페이지_요청(id, Role.OWNER);
-        오너_마이페이지_실패(response, errorMessage);
-    }
-
-    @DisplayName("오너회원탈퇴 성공")
-    @Test
-    void 오너회원탈퇴_성공() throws Exception {
-        final Long id = 1L;
-        Message message = new Message("회원탈퇴 성공");
-        willReturn(message).given(userService).withdraw(any(Long.class), any(Role.class));
-        final ResultActions response = 유저_회원탈퇴_요청(id, Role.OWNER);
-        오너_회원탈퇴_성공(response);
-    }
-
-    @DisplayName("오너회원탈퇴 실패")
-    @Test
-    void 오너회원탈퇴_실패() throws Exception {
-        final Long id = 1L;
-        Message errorMessage = new Message("회원탈퇴 실패");
-        willThrow(new NotFoundException("회원탈퇴 실패")).given(userService).withdraw(any(Long.class), any(Role.class));
-        final ResultActions response = 유저_회원탈퇴_요청(id, Role.OWNER);
-        오너_회원탈퇴_실패(errorMessage, response);
-    }
-
-    @DisplayName("오너 샵 조회 성공")
-    @Test
-    void 오너샵조회_성공() throws Exception {
-        List<OwnerShopResponseDto> ownerShopResponseDtoList = new ArrayList<>();
-        List<ShopTable> shopTables = new ArrayList<>();
-        Address address = new Address("인천시", "부평구", "산곡동");
-        Owner owner = Owner.builder()
-                .email("hjhj@naver.com")
-                .password("123")
-                .telNumber("123-123-123")
-                .name("사장")
-                .build();
-        shopTables.add(ShopTable.builder()
-                .maxUser(3)
-                .tableNumber(1)
-                .build());
-        shopTables.add(ShopTable.builder()
-                .maxUser(3)
-                .tableNumber(1)
-                .build());
-        Shop shop = Shop.builder()
-                .name("가게")
-                .address(address)
-                .owner(owner)
-                .telNumber("010-2222-1111")
-                .tableList(shopTables)
-                .build();
-        ownerShopResponseDtoList.add(new OwnerShopResponseDto(shop));
-        Long id = 1L;
-        willReturn(ownerShopResponseDtoList).given(userService).viewMyShop(any(Long.class));
-        ResultActions resultActions = 오너_샵조회_요청(id);
-        오너_샵조회_성공(resultActions, ownerShopResponseDtoList);
-    }
-
-    @DisplayName("오너 샵 조회 실패")
-    @Test
-    void 오너샵조회_실패() throws Exception {
-        Long id = 2L;
-        Message errorMessage = new Message("해당 유저 이메일이 존재하지 않습니다.");
-        willThrow(new NotFoundException("해당 유저 이메일이 존재하지 않습니다.")).given(userService).viewMyShop(any(Long.class));
-        final ResultActions response = 오너_샵조회_요청(id);
-        오너_샵조회_실패(response, errorMessage);
     }
 
     @DisplayName("즐겨찾기 등록 성공")
@@ -440,18 +253,18 @@ class UserControllerTest extends ApiDocument {
         List<UserVisitResponseDto> userVisitResponseList = new ArrayList<>();
         userVisitResponseList.add(new UserVisitResponseDto(shop, visitedDate));
         UserPageResponseDto userPageResponseDto = new UserPageResponseDto(user, userVisitResponseList);
-        willReturn(userPageResponseDto).given(userService).viewMypage(any(Long.class), any(Role.class));
-        ResultActions resultActions = 유저_마이페이지_조회_요청(id, Role.USER);
+        willReturn(userPageResponseDto).given(userService).viewMyPage(any(Long.class));
+        ResultActions resultActions = 유저_마이페이지_조회_요청(id);
         유저_마이페이지_조회_성공(resultActions, userPageResponseDto);
     }
 
-    @DisplayName("유저 방문매장 조회 실패")
+    @DisplayName("유저 마이페이지 조회 실패")
     @Test
     void 유저_마이페이지_조회_실패() throws Exception {
         Long id = 1L;
         Message errorMessage = new Message("해당 유저가 존재하지 않습니다.");
-        willThrow(new NotFoundException("해당 유저가 존재하지 않습니다.")).given(userService).viewMypage(any(Long.class), any(Role.class));
-        ResultActions resultActions = 유저_마이페이지_조회_요청(id, Role.USER);
+        willThrow(new NotFoundException("해당 유저가 존재하지 않습니다.")).given(userService).viewMyPage(any(Long.class));
+        ResultActions resultActions = 유저_마이페이지_조회_요청(id);
         유저_마이페이지_조회_실패(resultActions, errorMessage);
     }
 
@@ -469,8 +282,8 @@ class UserControllerTest extends ApiDocument {
                 .andDo(toDocument("user-view-page-fail"));
     }
 
-    private ResultActions 유저_마이페이지_조회_요청(Long id, Role role) throws Exception {
-        return mockMvc.perform(get("/api/v1/users/page/" + id + "/" + role));
+    private ResultActions 유저_마이페이지_조회_요청(Long id) throws Exception {
+        return mockMvc.perform(get("/api/v1/users/page/" + id));
     }
 
     private void 오너_샵조회_성공(ResultActions resultActions, List<OwnerShopResponseDto> ownerShopResponseDtoList) throws Exception {
@@ -480,25 +293,14 @@ class UserControllerTest extends ApiDocument {
                 .andDo(toDocument("owner-shop-view-success"));
     }
 
-    private void 오너_샵조회_실패(ResultActions resultActions, Message message) throws Exception {
-        resultActions.andExpect(status().isNotFound())
-                .andExpect(content().json(toJson(message)))
-                .andDo(print())
-                .andDo(toDocument("owner-shop-view-fail"));
-    }
-
-    private ResultActions 오너_샵조회_요청(Long id) throws Exception {
-        return mockMvc.perform(MockMvcRequestBuilders.get("/api/v1/users/shop/" + id));
-    }
-
     private ResultActions 유저_회원가입_요청(JoinRequestDto signUpRequest) throws Exception {
         return mockMvc.perform(MockMvcRequestBuilders.post("/api/v1/users/join")
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(toJson(signUpRequest)));
     }
 
-    private ResultActions 유저_회원탈퇴_요청(Long id, Role role) throws Exception {
-        return mockMvc.perform(delete("/api/v1/users/withdraw/" + id + "/" +role));
+    private ResultActions 유저_회원탈퇴_요청(Long id) throws Exception {
+        return mockMvc.perform(delete("/api/v1/users/withdraw/" + id));
     }
 
     private void 유저_회원가입_성공(Message message, ResultActions response) throws Exception {
@@ -515,20 +317,6 @@ class UserControllerTest extends ApiDocument {
                 .andDo(toDocument("user-signup-fail"));
     }
 
-    private void 오너_회원가입_성공(Message message, ResultActions response) throws Exception {
-        response.andExpect(status().isCreated())
-                .andExpect(content().json(toJson(message)))
-                .andDo(print())
-                .andDo(toDocument("owner-signup-success"));
-    }
-
-    private void 오너_회원가입_실패(Message errorMessage, ResultActions response) throws Exception {
-        response.andExpect(status().isBadRequest())
-                .andExpect(content().json(toJson(errorMessage)))
-                .andDo(print())
-                .andDo(toDocument("owner-signup-fail"));
-    }
-
     private void 유저_회원탈퇴_성공(ResultActions response) throws Exception {
         response.andExpect(status().isNoContent())
                 .andDo(print())
@@ -542,37 +330,11 @@ class UserControllerTest extends ApiDocument {
                 .andDo(toDocument("user-withdraw-fail"));
     }
 
-    private void 오너_회원탈퇴_성공(ResultActions response) throws Exception {
-        response.andExpect(status().isNoContent())
-                .andDo(print())
-                .andDo(toDocument("owner-withdraw-success"));
-    }
-
-    private void 오너_회원탈퇴_실패(Message errorMessage, ResultActions response) throws Exception {
-        response.andExpect(status().isNotFound())
-                .andExpect(content().json(toJson(errorMessage)))
-                .andDo(print())
-                .andDo(toDocument("owner-withdraw-fail"));
-    }
-
     private void 유저_로그아웃_실패(Message errorMessage, ResultActions resultActions) throws Exception {
         resultActions.andExpect(status().isNotFound())
                 .andExpect(content().json(toJson(errorMessage)))
                 .andDo(print())
                 .andDo(toDocument("user-logout-fail"));
-    }
-
-    private void 오너_로그아웃_성공(ResultActions resultActions) throws Exception {
-        resultActions.andExpect(status().isNoContent())
-                .andDo(print())
-                .andDo(toDocument("owner-logout-success"));
-    }
-
-    private void 오너_로그아웃_실패(Message errorMessage, ResultActions resultActions) throws Exception {
-        resultActions.andExpect(status().isNotFound())
-                .andExpect(content().json(toJson(errorMessage)))
-                .andDo(print())
-                .andDo(toDocument("owner-logout-fail"));
     }
 
     private void 유저_로그아웃_성공(ResultActions resultActions) throws Exception {
@@ -581,8 +343,8 @@ class UserControllerTest extends ApiDocument {
                 .andDo(toDocument("user-logout-success"));
     }
 
-    private ResultActions 유저_로그아웃_요청(Long id, Role role) throws Exception {
-        return mockMvc.perform(get("/api/v1/users/logout/" + id + "/" + role));
+    private ResultActions 유저_로그아웃_요청(Long id) throws Exception {
+        return mockMvc.perform(get("/api/v1/users/logout/" + id));
     }
 
     private void 유저_로그인_실패(Message errorMessage, ResultActions resultActions) throws Exception {
@@ -592,13 +354,6 @@ class UserControllerTest extends ApiDocument {
                 .andDo(toDocument("user-login-fail"));
     }
 
-    private void 오너_로그인_실패(Message errorMessage, ResultActions resultActions) throws Exception {
-        resultActions.andExpect(status().isNotFound())
-                .andExpect(content().json(toJson(errorMessage)))
-                .andDo(print())
-                .andDo(toDocument("owner-login-fail"));
-    }
-
     private void 유저_로그인_성공(LoginResponseDto loginResponseDto, ResultActions resultActions) throws Exception {
         resultActions.andExpect(status().isOk())
                 .andExpect(content().json(toJson(loginResponseDto)))
@@ -606,50 +361,11 @@ class UserControllerTest extends ApiDocument {
                 .andDo(toDocument("user-login-success"));
     }
 
-    private void 유저_마이페이지_성공(ResultActions response, UserPageResponseDto userPageResponseDto) throws Exception {
-        response.andExpect(status().isOk())
-                .andExpect(content().json(toJson(userPageResponseDto)))
-                .andDo(print())
-                .andDo(toDocument("user-view-page-success"));
-    }
-
-    private void 유저_마이페이지_실패(ResultActions response, Message errorMessage) throws Exception {
-        response.andExpect(status().isNotFound())
-                .andExpect(content().json(toJson(errorMessage)))
-                .andDo(print())
-                .andDo(toDocument("user-view-page-fail"));
-    }
-
-    private void 오너_마이페이지_성공(ResultActions response, UserPageResponseDto userPageResponseDto) throws Exception {
-        response.andExpect(status().isOk())
-                .andExpect(content().json(toJson(userPageResponseDto)))
-                .andDo(print())
-                .andDo(toDocument("owner-view-page-success"));
-    }
-
-    private void 오너_마이페이지_실패(ResultActions response, Message errorMessage) throws Exception {
-        response.andExpect(status().isNotFound())
-                .andExpect(content().json(toJson(errorMessage)))
-                .andDo(print())
-                .andDo(toDocument("owner-view-page-fail"));
-    }
-
-    private void 오너_로그인_성공(LoginResponseDto loginResponseDto, ResultActions resultActions) throws Exception {
-        resultActions.andExpect(status().isOk())
-                .andExpect(content().json(toJson(loginResponseDto)))
-                .andDo(print())
-                .andDo(toDocument("owner-login-success"));
-    }
-
     private ResultActions 유저_로그인_요청(LoginRequestDto LoginRequestDto) throws Exception {
         String content = objectMapper.writeValueAsString(LoginRequestDto);
         return mockMvc.perform(post("/api/v1/users/login")
                 .content(content)
                 .contentType(MediaType.APPLICATION_JSON));
-    }
-
-    private ResultActions 유저_마이페이지_요청(Long id, Role role) throws Exception {
-        return mockMvc.perform(MockMvcRequestBuilders.get("/api/v1/users/page/" + id + "/" + role));
     }
 
     private void 즐겨찾기_등록_성공(ResultActions response, Message message) throws Exception {
