+++ src/test/java/com/minionz/backend/user/service/UserServiceTest.java
@@ -5,12 +5,7 @@ import com.minionz.backend.common.domain.Message;
 import com.minionz.backend.common.exception.BadRequestException;
 import com.minionz.backend.common.exception.NotEqualsException;
 import com.minionz.backend.common.exception.NotFoundException;
-import com.minionz.backend.shop.controller.dto.ShopRequestDto;
-import com.minionz.backend.shop.controller.dto.ShopTableRequestDto;
-import com.minionz.backend.shop.service.ShopService;
 import com.minionz.backend.user.controller.dto.*;
-import com.minionz.backend.user.domain.Owner;
-import com.minionz.backend.user.domain.OwnerRepository;
 import com.minionz.backend.user.domain.User;
 import com.minionz.backend.user.domain.UserRepository;
 import org.junit.jupiter.api.AfterEach;
@@ -22,9 +17,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.web.WebAppConfiguration;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -39,21 +31,14 @@ public class UserServiceTest {
     private UserRepository userRepository;
 
     @Autowired
-    private OwnerRepository ownerRepository;
-
-    @Autowired
     private UserService userService;
 
     @Autowired
     private PasswordEncoder passwordEncoder;
 
-    @Autowired
-    private ShopService shopService;
-
     @AfterEach
     void cleanUp() {
         userRepository.deleteAll();
-        ownerRepository.deleteAll();
     }
 
     @Test
@@ -67,7 +52,6 @@ public class UserServiceTest {
                 .telNumber("11111111")
                 .password("1234")
                 .address(address)
-                .role(Role.USER)
                 .build();
         //when
         Message message = userService.signUp(joinRequestDto);
@@ -80,10 +64,10 @@ public class UserServiceTest {
         //given
         JoinRequestDto joinRequestDto = JoinRequestDto.builder()
                 .name("정재욱")
+                .nickName("바보")
                 .email("operation@naver.com")
                 .telNumber("11111111")
                 .password("1234")
-                .role(Role.OWNER)
                 .build();
         //when
         Message message = userService.signUp(joinRequestDto);
@@ -105,23 +89,7 @@ public class UserServiceTest {
                 .build();
         User save = userRepository.save(user);
         //when
-        Message message = userService.withdraw(save.getId(), Role.USER);
-        //then
-        assertThat(message.getMessage()).isEqualTo("회원탈퇴 성공");
-    }
-
-    @Test
-    void 회원탈퇴_성공_테스트_오너() {
-        //given
-        Owner owner = Owner.builder()
-                .email("wodnr8462@naver.com")
-                .name("정재욱")
-                .password("1234")
-                .telNumber("1111111")
-                .build();
-        Owner save = ownerRepository.save(owner);
-        //when
-        Message message = userService.withdraw(save.getId(), Role.OWNER);
+        Message message = userService.withdraw(save.getId());
         //then
         assertThat(message.getMessage()).isEqualTo("회원탈퇴 성공");
     }
@@ -145,7 +113,6 @@ public class UserServiceTest {
                 .nickName("라이언")
                 .telNumber("11111111")
                 .password("1234")
-                .role(Role.USER)
                 .address(address)
                 .build();
         //when
@@ -165,36 +132,15 @@ public class UserServiceTest {
                 .telNumber("11111111")
                 .password("1234")
                 .address(address)
-                .role(Role.USER)
                 .build();
         userService.signUp(joinRequestDto);
         User findUser = userRepository.findByEmail("operation@naver.com")
                 .orElseThrow(() -> new NotFoundException("회원가입 실패"));
-        LoginRequestDto LoginRequestDto = new LoginRequestDto("operation@naver.com", "1234", Role.USER);
+        LoginRequestDto LoginRequestDto = new LoginRequestDto("operation@naver.com", "1234");
         //when
         LoginResponseDto login = userService.login(LoginRequestDto);
         //then
         assertThat(login.getId()).isEqualTo(findUser.getId());
-    }
-
-    @Test
-    void 로그인_성공_테스트_오너() {
-        //given
-        JoinRequestDto joinRequestDto = JoinRequestDto.builder()
-                .name("정재욱")
-                .email("operation@naver.com")
-                .telNumber("11111111")
-                .password("1234")
-                .role(Role.OWNER)
-                .build();
-        userService.signUp(joinRequestDto);
-        Owner findOwner = ownerRepository.findByEmail("operation@naver.com")
-                .orElseThrow(() -> new NotFoundException("회원가입 실패"));
-        LoginRequestDto LoginRequestDto = new LoginRequestDto("operation@naver.com", "1234", Role.OWNER);
-        //when
-        LoginResponseDto login = userService.login(LoginRequestDto);
-        //then
-        assertThat(login.getId()).isEqualTo(findOwner.getId());
     }
 
     @Test
@@ -207,7 +153,7 @@ public class UserServiceTest {
                 .nickName("donglee99")
                 .telNumber("010111111111")
                 .build();
-        LoginRequestDto LoginRequestDto = new LoginRequestDto("jh3j741@naver.com", "123456", Role.USER);
+        LoginRequestDto LoginRequestDto = new LoginRequestDto("jh3j741@naver.com", "123456");
         userRepository.save(user);
         //when
         //then
@@ -226,7 +172,7 @@ public class UserServiceTest {
                 .nickName("donglee99")
                 .telNumber("010111111111")
                 .build();
-        LoginRequestDto LoginRequestDto = new LoginRequestDto("jhnj741@naver.com", "12346", Role.USER);
+        LoginRequestDto LoginRequestDto = new LoginRequestDto("jhnj741@naver.com", "12346");
         userRepository.save(user);
         //when
         //then
@@ -247,9 +193,9 @@ public class UserServiceTest {
                 .telNumber("010111111111")
                 .address(address)
                 .build();
-        userRepository.save(user);
+        User save = userRepository.save(user);
         //when
-        UserPageResponseDto userPageResponseDto = userService.viewMypage(user.getId(), Role.USER);
+        UserPageResponseDto userPageResponseDto = userService.viewMyPage(save.getId());
         //then
         assertThat(userPageResponseDto.getNickname()).isEqualTo("donglee99");
         assertThat(userPageResponseDto.getTelNumber()).isEqualTo("010111111111");
@@ -268,7 +214,7 @@ public class UserServiceTest {
         userRepository.save(user);
         //when
         //then
-        assertThatThrownBy(() -> userService.viewMypage(2L, Role.USER))
+        assertThatThrownBy(() -> userService.viewMyPage(2L))
                 .isInstanceOf(NotFoundException.class)
                 .hasMessage("해당 유저 이메일이 존재하지 않습니다.");
     }
@@ -286,7 +232,6 @@ public class UserServiceTest {
                 .telNumber("11111111")
                 .password("12345678")
                 .address(address)
-                .role(Role.USER)
                 .build();
         userService.signUp(joinRequestDto);
         User user = userRepository.findByEmail("operation@naver.com")
@@ -298,63 +243,5 @@ public class UserServiceTest {
                 () -> assertNotEquals(rawPassword, encode),
                 () -> assertTrue(passwordEncoder.matches(rawPassword, encode))
         );
-    }
-
-    @Test
-    void 오너샵조회_성공() {
-        //given
-        Address address = new Address("믿음", "소망", "씨티");
-        List<ShopTableRequestDto> shopTables1 = new ArrayList<>();
-        shopTables1.add(new ShopTableRequestDto(2));
-        shopTables1.add(new ShopTableRequestDto(4));
-        shopTables1.add(new ShopTableRequestDto(4));
-        List<ShopTableRequestDto> shopTables2 = new ArrayList<>();
-        shopTables2.add(new ShopTableRequestDto(4));
-        shopTables2.add(new ShopTableRequestDto(4));
-        shopTables2.add(new ShopTableRequestDto(4));
-        Owner owner = Owner.builder()
-                .name("주인")
-                .email("jhnj841@naba.com")
-                .password("123")
-                .telNumber("123123")
-                .build();
-        Owner savedOwner = ownerRepository.save(owner);
-        ShopRequestDto shopRequestDto1 = new ShopRequestDto("맘스터치1", address, "010-111-33332", shopTables1, savedOwner.getId());
-        ShopRequestDto shopRequestDto2 = new ShopRequestDto("맘스터치2", address, "010-111-33333", shopTables2, savedOwner.getId());
-        shopService.save(shopRequestDto1);
-        shopService.save(shopRequestDto2);
-        //when
-        List<OwnerShopResponseDto> ownerShopResponseDtoList = userService.viewMyShop(savedOwner.getId());
-        //then
-        assertThat(ownerShopResponseDtoList.size()).isEqualTo(2);
-    }
-
-    @Test
-    void 오너샵조회_실패() {
-        //given
-        Address address = new Address("믿음", "소망", "씨티");
-        List<ShopTableRequestDto> shopTables1 = new ArrayList<>();
-        shopTables1.add(new ShopTableRequestDto(2));
-        shopTables1.add(new ShopTableRequestDto(4));
-        shopTables1.add(new ShopTableRequestDto(4));
-        List<ShopTableRequestDto> shopTables2 = new ArrayList<>();
-        shopTables2.add(new ShopTableRequestDto(4));
-        shopTables2.add(new ShopTableRequestDto(4));
-        shopTables2.add(new ShopTableRequestDto(4));
-        Owner owner = Owner.builder()
-                .name("주인")
-                .email("jhnj841@naba.com")
-                .password("123")
-                .telNumber("123123")
-                .build();
-        Owner savedOwner = ownerRepository.save(owner);
-        ShopRequestDto shopRequestDto1 = new ShopRequestDto("맘스터치1", address, "010-111-33332", shopTables1, savedOwner.getId());
-        ShopRequestDto shopRequestDto2 = new ShopRequestDto("맘스터치2", address, "010-111-33333", shopTables2, savedOwner.getId());
-        shopService.save(shopRequestDto1);
-        shopService.save(shopRequestDto2);
-        //when
-        //then
-        assertThatThrownBy(() -> userService.viewMyShop(2L))
-                .isInstanceOf(NotFoundException.class);
     }
 }
